requirements.txt
rumps>=0.4.0
pynput>=1.7.6
mediapipe>=0.10.0
opencv-python>=4.8.0
numpy>=1.24.0
pyobjc-framework-Quartz>=9.0
websockets>=12.0
sounddevice>=0.5.0
